Add comments to Parcel::ParseFrom method
https://gitcode.com/openharmony/commonlibrary_c_utils/issues/525

ParseFrom assigns the given pointer to the internal member data_.
The Parcel object takes ownership of this memory, and it will be freed
automatically when the Parcel is destroyed. Do not manually free or reuse
the original pointer after calling ParseFrom, otherwise double-free or
use-after-free errors may occur.

To note developers using this method more correctly.

Signed-off-by: xingzeng <xingzeng2@huawei.com>